经验分享模块的部分开发

+++ src/main/java/com/xmut/xmut_java/sys/entity/SysUserAndExperience.java
@@ -7,8 +7,8 @@ import com.xmut.xmut_java.common.BaseEntity;
 public class SysUserAndExperience extends BaseEntity{
 	@TableId(value="id", type=IdType.AUTO)
 	private Long id;
-	private int userId;
-	private int experienceId;
+	private Long userId;
+	private Long experienceId;
 	
 	public Long getId() {
 		return id;
@@ -16,16 +16,16 @@ public class SysUserAndExperience extends BaseEntity{
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public int getExperienceId() {
+	public Long getExperienceId() {
 		return experienceId;
 	}
-	public void setExperienceId(int experienceId) {
+	public void setExperienceId(Long experienceId) {
 		this.experienceId = experienceId;
 	}
 }
